# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: REST API for managing bank cards with JWT authentication
  version: 1.0.0
  contact:
    name: Frolov Nikita Macsimovich
    email: nikitka_frolov_2014@inbox.ru
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/cards:
    get:
      tags: [Cards]
      summary: "Get user's cards"
      description: "Get paginated list of user's bank cards"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "owner"
      responses:
        '200':
          description: "List of user's cards"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponse'
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/cards/by-number:
    get:
      tags: [Cards]
      summary: "Get card by number"
      description: "Get card details by card number (original number)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
            example: "4111111111111111"
      responses:
        '200':
          description: "Card details"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: "Card not found"
  /api/cards/balance:
    get:
      tags: [Cards]
      summary: "Get card balance"
      description: "Get balance of specific card"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card balance"
          content:
            application/json:
              schema:
                type: number
                format: double
                example: 1500.50
        '404':
          description: "Card not found"
  /api/cards/activate:
    put:
      tags: [Cards]
      summary: "Activate card"
      description: "Activate user's card"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card activated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
  /api/cards/search:
    get:
      tags: [Cards]
      summary: "Search cards by owner name"
      description: "Search user's cards by owner name (case insensitive)"
      security:
        - bearerAuth: []
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
            example: "IVAN"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "Search results"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponse'

  # ========== TRANSACTION ENDPOINTS ==========
...
  /api/admin/cards/block:
    put:
      tags: [Admin]
      summary: "Block any card (Admin only)"
      description: "Block any card in system (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card blocked successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/activate:
    put:
      tags: [Admin]
      summary: "Activate any card (Admin only)"
      description: "Activate any card in system (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card activated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
...
  /api/admin/cards/balance:
    put:
      tags: [Admin]
      summary: "Update card balance (Admin only)"
      description: "Update card balance (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
        - name: newBalance
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: 0
            example: 5000.00
      responses:
        '200':
          description: "Balance updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
...
  /api/cards/block-request:
    post:
      tags: [ Cards ]
      summary: "Request card block"
      description: "Request card blocking (creates pending request for admin approval)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
            example: "4111111111111111"
        - name: reason
          in: query
          required: false
          schema:
            type: string
            example: "Card lost"
      responses:
        '200':
          description: "Block request created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'
        '400':
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/block-requests:
    get:
      tags: [ Cards ]
      summary: "Get user's block requests"
      description: "Get paginated list of user's block requests"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "List of block requests"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestPageResponse'
...
  /api/cards/block:
    put:
      tags: [ Cards ]
      summary: "Block card immediately (Admin only)"
      description: "Block card immediately without approval workflow (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
            example: "4111111111111111"
      responses:
        '200':
          description: "Card blocked successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: "Admin access required"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CardResponse:
      type: object
      properties:
        maskedNumber:
          type: string
          example: "**** **** **** 1111"
        owner:
          type: string
          example: "IVAN IVANOV"
        expiryDate:
          type: string
          format: date
          example: "2026-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          example: "ACTIVE"
        balance:
          type: number
          format: double
          example: 1500.50
        userId:
          type: integer
          example: 2
    AdminCardResponse:
      type: object
      properties:
        maskedNumber:
          type: string
          example: "**** **** **** 1111"
        owner:
          type: string
          example: "IVAN IVANOV"
        expiryDate:
          type: string
          format: date
          example: "2026-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          example: "ACTIVE"
        balance:
          type: number
          format: double
          example: 1500.50
        userId:
          type: integer
          example: 2
        username:
          type: string
          example: "user1"
    CardPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalPages:
          type: integer
          example: 5
        totalElements:
          type: integer
          example: 42
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        last:
          type: boolean
          example: false
        first:
          type: boolean
          example: true
        numberOfElements:
          type: integer
          example: 10
        empty:
          type: boolean
          example: false
...
    BlockRequestResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cardNumber:
          type: string
          description: "Encrypted card number"
          example: "Nj0RRQL9mvNhBh7DfNHbOHrVbNFf2gXt/a8eAaX1R8E="
        cardMasked:
          type: string
          example: "**** **** **** 1111"
        owner:
          type: string
          example: "IVAN IVANOV"
        userId:
          type: integer
          example: 2
        username:
          type: string
          example: "user1"
        requestDate:
          type: string
          format: date-time
          example: "2024-01-15T14:30:00"
        reason:
          type: string
          example: "Card lost"
        status:
          type: string
          enum: [ PENDING, APPROVED, REJECTED ]
          example: "PENDING"
        processedDate:
          type: string
          format: date-time
          example: "2024-01-15T15:30:00"
        processedByUsername:
          type: string
          example: "admin"

      BlockRequestPageResponse:
        type: object
        properties:
          content:
            type: array
            items:
              $ref: '#/components/schemas/BlockRequestResponse'
          totalPages:
            type: integer
            example: 5
          totalElements:
            type: integer
            example: 42
          size:
            type: integer
            example: 10
          number:
            type: integer
            example: 0
          last:
            type: boolean
            example: false
          first:
            type: boolean
            example: true
          numberOfElements:
            type: integer
            example: 10
          empty:
            type: boolean
            example: false

    # ========== TRANSACTION SCHEMAS ==========
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T14:30:00"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input parameters"
        path:
          type: string
          example: "/api/cards/by-number"
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
          example: "cardNumber"
        message:
          type: string
          example: "Card number must be 16 digits"
        rejectedValue:
          type: string
          example: "123"
security:
  - bearerAuth: []